requirements.txt
streamlit==1.31.0
mysql-connector-python==8.2.0
pandas==2.3.3
plotly==5.18.0
reportlab==4.0.8
python-dotenv==1.0.0